more tests and refactorint

+++ Betting/Config/ConfigManagerInterface.cs
@@ -18,11 +18,11 @@ namespace Betting.Config
         abstract public int GetDrawMixedMargin();
         abstract public int GetMatchDay();
         abstract public int GetReverseDays();
-        abstract public float GetMinOdds();
-        abstract public float GetMaxOdds();
-        abstract public float GetMinMetricCorrect();
-        abstract public float GetMinYearProfit();
-        abstract public float GetSuccessRate();
+        abstract public double GetMinOdds();
+        abstract public double GetMaxOdds();
+        abstract public double GetMinMetricCorrect();
+        abstract public double GetMinYearProfit();
+        abstract public double GetSuccessRate();
         abstract public int GetFilterTopRate();
         abstract public int GetFilterTopProfit();
         abstract public string GetBetStyle();
